# Month3Lesson3/api/docs/swagger.yaml
definitions:
  models.Basket:
    properties:
      customer_id:
        type: string
      id:
        type: string
      total_sum:
        type: string
    type: object
  models.BasketProduct:
    properties:
      basket_id:
        type: string
      id:
        type: string
      product_id:
        type: string
      quantity:
        type: integer
    type: object
  models.BasketResponse:
    properties:
      baskets:
        items:
          $ref: '#/definitions/models.Basket'
        type: array
      count:
        type: integer
    type: object
...
  models.CreateBasket:
    properties:
      customer_id:
        type: string
      total_sum:
        type: integer
    type: object
  models.CreateBasketProduct:
    properties:
      basket_id:
        type: string
      product_id:
        type: string
      quantity:
        type: integer
    type: object
...
  models.Response:
    properties:
      data: {}
      description:
        type: string
      statusCode:
        type: integer
    type: object
...
  models.UpdateBasket:
    properties:
      customer_id:
        type: string
      id:
        type: string
      total_sum:
        type: integer
    type: object
  models.UpdateBasketProduct:
    properties:
      basket_id:
        type: string
      id:
        type: string
      product_id:
        type: string
      quantity:
        type: integer
    type: object
...
info:
  contact: {}
  description: This is a sample server celler server
  title: Swagger Example API
  version: "1.0"
paths:
  /basket:
    get:
      consumes:
      - application/json
      description: Get basket list
      parameters:
      - description: page
        in: query
        name: page
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BasketResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get basket list
      tags:
      - basket
    post:
      consumes:
      - application/json
      description: create a new basket
      parameters:
      - description: basket
        in: body
        name: basket
        required: true
        schema:
          $ref: '#/definitions/models.CreateBasket'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Basket'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Creates a new basket
      tags:
      - basket
  /basket/{id}:
    delete:
      consumes:
      - application/json
      description: Delete basket
      parameters:
      - description: basket
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Delete basket
      tags:
      - basket
    get:
      consumes:
      - application/json
      description: Get basket by id
      parameters:
      - description: basket
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Basket'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Gets basket
      tags:
      - basket
    put:
      consumes:
      - application/json
      description: Update basket
      parameters:
      - description: basket_id
        in: path
        name: id
        required: true
        type: string
      - description: basket
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UpdateBasket'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Basket'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Update basket
      tags:
      - basket
  /basketProduct:
    get:
      consumes:
      - application/json
      description: Get basketProduct list
      parameters:
      - description: page
        in: query
        name: page
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BasketProduct'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get basketProduct list
      tags:
      - basketProduct
    post:
      consumes:
      - application/json
      description: create a new basket product
      parameters:
      - description: basketProduct
        in: body
        name: basketProduct
        required: true
        schema:
          $ref: '#/definitions/models.CreateBasketProduct'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.BasketProduct'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Creates a new basket product
      tags:
      - basketProduct
  /basketProduct/{id}:
    delete:
      consumes:
      - application/json
      description: Delete basket product
      parameters:
      - description: basket_product_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Delete basket product
      tags:
      - basketProduct
    get:
      consumes:
      - application/json
      description: Get basketProduct by id
      parameters:
      - description: basketProduct
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BasketProduct'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Gets basketProduct
      tags:
      - basketProduct
    put:
      consumes:
      - application/json
      description: Update basket product
      parameters:
      - description: basketProduct_id
        in: path
        name: id
        required: true
        type: string
      - description: basketProduct
        in: body
        name: basketProduct
        required: true
        schema:
          $ref: '#/definitions/models.UpdateBasketProduct'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.BasketProduct'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Update basket product
      tags:
      - basketProduct
...
swagger: "2.0"